feat: ClassItem 컴포넌트 분리 및 삭제 버튼 추가
